Regression in Sodium.java fixed

+++ src/main/java/com/goterl/lazycode/lazysodium/Sodium.java
@@ -119,9 +119,9 @@ public class Sodium {
     //// -------------------------------------------|
 
     public native int crypto_pwhash(byte[] outputHash,
-                                 int outputHashLen,
+                                 long outputHashLen,
                                  byte[] password,
-                                 int passwordLen,
+                                 long passwordLen,
                                  byte[] salt,
                                  long opsLimit,
                                  int memLimit,
@@ -129,11 +129,11 @@ public class Sodium {
 
     public native int crypto_pwhash_str(byte[] outputStr,
                                      byte[] password,
-                                     int passwordLen,
+                                     long passwordLen,
                                      long opsLimit,
                                      int memLimit);
 
-    public native int crypto_pwhash_str_verify(byte[] hash, byte[] password, int passwordLen);
+    public native int crypto_pwhash_str_verify(byte[] hash, byte[] password, long passwordLen);
 
     public native int crypto_pwhash_str_needs_rehash(byte[] hash, long opsLimit, int memLimit);
